Add initial project setup and requirements
Added vcs.xml for Git VCS directory mapping and a README file that includes a brief description and installation process for the project. Also included a requirements.txt file which contains all the necessary packages for setting up an NLP model with a Flask front end. This setup enables easy replication and deployment of the project.