Add data sampling and file information to file viewer
This commit tweaks the 'view_file_contents' route and template to provide additional information to the user and reduce the potential load on the application. In the previous implementation, the entirety of a file was loaded into a dataframe and displayed. Now, a random sampling of 10 rows is shown instead, significantly reducing the viewing load, and displaying the total number of rows in the file for user understanding.

Additionally, the file's name is shown above the sample, giving the user context of the data displayed. This small modification provides a breather to our application while keeping users informed enough to guide their data processing choices.