Update app base title and UI strings for clarity
The commit refreshes the base title in 'base.html' to 'NLP Modeling Tool' for better relevance to the application's purpose, and also updates link and button texts in 'view_worksheet.html' and 'view_file_contents.html' for clarity. Missing values representation is changed to a ratio in 'view_worksheet.html', and the sample size for file contents view is now reduced to 5 rows for better manageability. The unnecessary CSV file name user input section has been removed from 'view_worksheet.html' as the processed_file_name has been standardized. Changes in 'routes.py' reflect the corresponding updates in the templates as well as optimize the file processing flow. This commit aligns the app UI better with the added excel support and the enhanced data processing workflow.