"Add structure for Flask application"
This commit sets up the basic structure of a Flask application, including blueprints and routing for the main app. It includes creating new routes in "app/routes.py", as well as the addition of a run file "run.py". Furthermore, new HTML templates for 'index', 'data management', 'data modeling', and 'reporting' pages were added in "app/templates". Configuration file for the Flask app is also added as part of this commit. These changes were necessary to get the foundational setup done for the application.