Improve data processing output handling
Added ability to save preprocessed data to a .csv file. This change was made in the back-end logic in the 'data_modeling.py' script, allowing users to specify an output filename for their processed data. If the given filename does not end in '.csv', the extension is added, and potential file overwriting is prevented with an existence check. To accommodate these changes, the front-end (template and PreprocessingForm) was updated to include a field to input the output filename. This promotes data conservation and improves user control over data management.