Add Kaggle reference to README file
A reference to a Kaggle data source has been added to the README file. This reference provides an additional demonstration dataset for users interested in more extensive testing or experimentation. The dataset contains Amazon reviews, offering a concrete and relatable context for users to better understand the capabilities of the tool.