Add processed_review_text to test_file.csv
A processed_text column has been added to the test_file.csv file under app/data/data_processed. The purpose of this addition is to have a source of preprocessed review texts available for analysis. Preprocessing involves cleaning the text of noise, normalizing it, and breaking it down into its constituent words (tokens), which are represented in an array.