"Replace 'reporting' module with 'data_preprocessing'
The 'reporting' module was causing confusion due to the naming. The term 'reporting' was not accurately describing the function of this module. Therefore, it was decided to rename it to 'data_preprocessing', which more accurately depicts the module's function of preparing data for the model. This also explains why many changes in routing and templates were required. Now, all the references to 'reporting' have been replaced with 'data_preprocessing' and everything should work as expected."

Note: This message has been edited to fit the format and character limit, the original message was quite detailed and well-written, but was too long.