Remove processed test data
Removed test_file.csv from data_processed folder. The file contained processed test data that is no longer needed in the production environment, as it unnecessarily increases the size of the codebase. Additionally, it's generally not best practice to include data in source control if it can be regenerated from the application itself. The changes in the data were the result of test scenarios, thus bringing