Add display_data route and template in model_builder
This commit introduces a new route, display_data, in the model_builder blueprint which allows the preview of selected data from a model building form. It also includes a new template, display_data.html, to display this data.

The context for the template is provided by the 'display_data' function in the 'routes.py' file where the selected data is retrieved and processed to HTML. If no columns or specific columns are selected from the form, the function then redirects to this template to preview the data.

In case of any errors, flash messages are used to provide necessary feedback to the user. Various checks have been added to prevent issues in scenarios where no file or specific column is selected.