Add data filtering functionality to file manager
This update adds a feature to filter data within the file manager. It includes creating new routes for initiating and performing filtering and for saving filtered data. New HTML templates for these specific steps are included as well. A filter data button has been added to the file listing in the file manager.